split learns cartpole to separate test
